backend/requirements.txt
Django==2.2.5
djangorestframework==3.10.2
Markdown==3.1.1
django-filter==2.2.0
django-allauth==0.40.0
graphene-django==2.5.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.1
celery==4.3.0
redis==3.3.8
